+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useCaseBlocksLogic/utils/create-kyc-blocks.tsx
@@ -1,28 +1,60 @@
-import { TWorkflowById } from '@/domains/workflows/fetchers';
 import { KycBlock } from '@/lib/blocks/components/KycBlock/KycBlock';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
+import { ComponentProps } from 'react';
 
-export const createKycBlocks = (workflow: TWorkflowById) => {
+export const createKycBlocks = (individuals: ComponentProps<typeof KycBlock>[]) => {
   const blocks = createBlocksTyped().addBlock();
 
-  const childWorkflows = workflow?.childWorkflows?.filter(
-    childWorkflow => childWorkflow?.context?.entity?.type === 'individual',
+  if (!individuals?.length) {
+    return [];
+  }
+
+  individuals.forEach(
+    ({
+      documents,
+      entityData,
+      status,
+      kycSession,
+      aml,
+      isActionsDisabled,
+      isLoadingReuploadNeeded,
+      isLoadingApprove,
+      onInitiateKyc,
+      onInitiateSanctionsScreening,
+      onApprove,
+      onReuploadNeeded,
+      reasons,
+      isReuploadNeededDisabled,
+      isApproveDisabled,
+      isInitiateKycDisabled,
+      isInitiateSanctionsScreeningDisabled,
+    }) => {
+      blocks.addCell({
+        type: 'node',
+        value: (
+          <KycBlock
+            documents={documents}
+            entityData={entityData}
+            status={status}
+            kycSession={kycSession}
+            aml={aml}
+            isActionsDisabled={isActionsDisabled}
+            isLoadingReuploadNeeded={isLoadingReuploadNeeded}
+            isLoadingApprove={isLoadingApprove}
+            onInitiateKyc={onInitiateKyc}
+            onInitiateSanctionsScreening={onInitiateSanctionsScreening}
+            onApprove={onApprove}
+            onReuploadNeeded={onReuploadNeeded}
+            reasons={reasons}
+            isReuploadNeededDisabled={isReuploadNeededDisabled}
+            isApproveDisabled={isApproveDisabled}
+            isInitiateKycDisabled={isInitiateKycDisabled}
+            isInitiateSanctionsScreeningDisabled={isInitiateSanctionsScreeningDisabled}
+          />
+        ),
+      });
+    },
   );
-
-  if (!childWorkflows?.length) return [];
-
-  childWorkflows.forEach(childWorkflow => {
-    blocks.addCell({
-      type: 'node',
-      value: (
-        <KycBlock
-          parentWorkflowId={workflow.id}
-          childWorkflow={childWorkflow}
-          key={childWorkflow?.id}
-        />
-      ),
-    });
-  });
 
   return blocks.build();
 };