+++ apps/backoffice-v2/src/pages/MerchantMonitoring/schemas.ts
@@ -125,9 +125,13 @@ export const MerchantMonitoringSearchSchema = BaseSearchSchema.extend({
     .catch('All'),
   from: z.string().date().optional(),
   to: z.string().date().optional(),
+  allowAllDates: z
+    .string()
+    .transform(value => value === 'true')
+    .optional(),
   isCreating: z
     .string()
-    .transform(value => (value === 'true' ? true : false))
+    .transform(value => value === 'true')
     .optional(),
 });