+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ballerine/workflow-core
 
+## 0.6.114
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.91
+
+## 0.6.113
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.90
+
+## 0.6.112
+
+### Patch Changes
+
+- Bump
+
+## 0.6.111
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.89
+- bump
+
 ## 0.6.110
 
 ### Patch Changes